+++ LIBFT/ft_substr.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_substr.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: bokim <bokim@student.42.fr>                +#+  +:+       +#+        */
+/*   By: boyoung <boyoung@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/10 13:41:25 by bokim             #+#    #+#             */
-/*   Updated: 2025/11/13 14:59:51 by bokim            ###   ########.fr       */
+/*   Updated: 2025/11/19 12:34:23 by boyoung          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ char	*ft_substr(char const *s, unsigned int start, size_t len)
 	res = malloc((ft_count_len(s, start, len) + 1) * sizeof(char));
 	if (!res)
 		return (0);
-	if (start < (unsigned int)ft_strlen(s))
+	if ((size_t)start < ft_strlen(s))
 	{
 		i = start;
 		while (s[i] && j < len)
@@ -37,17 +37,17 @@ char	*ft_substr(char const *s, unsigned int start, size_t len)
 
 static int	ft_count_len(char const *s, unsigned int start, size_t len)
 {
-	int				count;
-	unsigned int	i;
+	size_t	count;
+	size_t	i;
 
 	count = 0;
-	i = (unsigned int)ft_strlen(s);
-	if (start < i)
+	i = ft_strlen(s);
+	if ((size_t)start < i)
 	{
-		if (len < i - start)
+		if (len < i - (size_t)start)
 			count = len;
 		else
-			count = i - start;
+			count = i - (size_t)start;
 	}
 	return (count);
 }
